Changed exception class raised on error.
According to the Django docs, errors in sending mail should raise an
smtplib.SMTPException.

Additionally, missing settings now raises an ImproperlyConfigured error
and ignores fail_silently, as this isn't a problem sending mail but is a
configuration error.

According to the Django docs, if fail_silently is False, send_mail
will raise a SMTPException, which are all problems either communicating
with or problems with the remote server processing the request. I
interpret this to be intended to silence problems out of the
application's control, such as the mail server being down or some
parameters or a particular message being invalid. A configuration error
on the client side doesn't fall into this category of error and
therefor should not obey fail_silently.