Added unicode tests and updated test cases
This was a big change mostly because the tests do a lot of manual work
to decode the requests in order to verify them, and that necessarily
involves some tricky encoding work.

Good news is that unicode is now supported in both the email body and
the subject, in both Python 2.7 and 3.4!